--- lib/busted/middleware/busted_tracer.rb
@@ -1,26 +1,0 @@
-require "logger"
-
-class BustedTracer
-  def initialize(app)
-    @app = app
-  end
-
-  def call(env)
-    Busted.start
-    request = @app.call(env)
-    report = Busted.finish
-
-    logger.info "[Cache Invalidations] methods=#{report[:invalidations][:method]} constants=#{report[:invalidations][:constant]}"
-    request
-  end
-
-  private
-
-  def logger
-    if defined?(Rails)
-      Rails.logger
-    else
-      Logger.new(STDOUT)
-    end
-  end
-end
